feat(dashboard,core): effectiveness metrics panel (Workstream K2)
- Move computeMetricsFromDisk to @murmurations-ai/core (packages/core/src/metrics/disk-snapshot.ts)
  so the dashboard can reuse it without duplicating the JSONL reader.
- CLI's metrics-cmd.ts becomes a thin formatter wrapper around the shared compute.
- Add readMetricsSnapshot to dashboard-tui/data.ts (30-day default window).
- Add renderMetrics panel between Governance and Activity panels:
  - Aggregate: wakes / completion rate / spend (color-graded)
  - Per-accountability met-rate (top 6, color-graded by ≥80/≥50/red)
  - Falls back to gentle hint when no observations yet

The panel renders the same numbers as `murmuration metrics --json` so
dashboard and CLI stay coherent. GitHub-derived metrics (closure rate,
age distribution) are still skipped here — they land with the
metrics.snapshot daemon RPC follow-up.
