+++ packages/core/src/index.ts
@@ -61,3 +61,4 @@ export * from "./daemon/command-executor.js";
 export * from "./done-criteria/index.js";
 export * from "./metrics/effectiveness.js";
 export * from "./metrics/observations.js";
+export * from "./metrics/disk-snapshot.js";
